requirements.txt
Flask
gunicorn
psycopg2-binary
mysql-connector-python
python-dotenv